+++ src/proxy/src/storage/storage_api.rs
@@ -1,10 +1,10 @@
 use candid::Principal;
 use canister_types::models::{
     api_error::ApiError, attendee::Attendee, boosted::Boost, event::Event,
-    event_collection::EventCollection, friend_request::FriendRequest, group::Group, log::Logger,
-    member::Member, member_collection::MemberCollection, notification::Notification,
-    profile::Profile, report::Report, reward::RewardableActivity,
-    user_notifications::UserNotifications,
+    event_collection::EventCollection, friend_request::FriendRequest, group::Group,
+    group_transfer_request::GroupTransferRequest, log::Logger, member::Member,
+    member_collection::MemberCollection, notification::Notification, profile::Profile,
+    report::Report, reward::RewardableActivity, user_notifications::UserNotifications,
 };
 use ic_stable_structures::{
     memory_manager::{MemoryId, MemoryManager, VirtualMemory},
@@ -54,6 +54,8 @@ pub static IDS_MEMORY_ID: MemoryId = MemoryId::new(19);
 
 pub static REWARD_BUFFER_MEMORY_ID: MemoryId = MemoryId::new(20);
 pub static REWARD_CANISTER_MEMORY_ID: MemoryId = MemoryId::new(21);
+
+pub static GROUP_TRANSFER_REQUESTS_MEMORY_ID: MemoryId = MemoryId::new(22);
 
 /// A reference to a `StableBTreeMap` that is wrapped in a `RefCell`.
 ///# Generics
@@ -352,6 +354,10 @@ thread_local! {
 
     pub static IDS: StorageRef<String, u64> = RefCell::new(
         StableBTreeMap::init(MEMORY_MANAGER.with(|p| p.borrow().get(IDS_MEMORY_ID)))
+    );
+
+    pub static GROUP_TRANSFER_REQUESTS: StorageRef<u64, GroupTransferRequest> = RefCell::new(
+        StableBTreeMap::init(MEMORY_MANAGER.with(|p| p.borrow().get(GROUP_TRANSFER_REQUESTS_MEMORY_ID)))
     );
 
 }